fixing incorrect --CLEAN-- sections
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/branches/PHP_5_3@278816 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/standard/tests/file/file_get_contents_error.phpt
@@ -38,9 +38,12 @@ echo "\n*** Done ***\n";
 --CLEAN--
 <?php
 $file_path = dirname(__FILE__);
-unlink($file_path."/file_put_contents_error.tmp");
-unlink($file_path."/file_put_contents1.tmp");
-
+if(file_exists($file_path."/file_put_contents_error.tmp")) {
+  unlink($file_path."/file_put_contents_error.tmp");
+}
+if(file_exists($file_path."/file_put_contents1.tmp")) {
+  unlink($file_path."/file_put_contents1.tmp");
+}
 ?>
 --EXPECTF--
 *** Testing error conditions ***
