+++ ext/standard/tests/file/is_file_error.phpt
@@ -29,8 +29,12 @@ echo "\n*** Done ***";
 --CLEAN--
 <?php
 $file_path = dirname(__FILE__);
-unlink($file_path."/is_file_error.tmp");
-unlink($file_path."/is_file_error1.tmp");
+if(file_exists($file_path."/is_file_error.tmp")) {
+  unlink($file_path."/is_file_error.tmp");
+}
+if(file_exists($file_path."/is_file_error1.tmp")) {
+  unlink($file_path."/is_file_error1.tmp");
+}
 ?>
 --EXPECTF--
 *** Testing is_file() error conditions ***
